user_service/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
psycopg2-binary==2.9.3
Werkzeug==2.2.2
pulsar-client==3.3.0